server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.120.2",
    "langchain>=1.0.2",
    "langchain-openai>=1.0.1",
    "langchain-tavily>=0.2.12",
    "langchain-text-splitters>=1.0.0",
    "langgraph-checkpoint-postgres>=3.0.0",
    "langgraph-cli[inmem]>=0.4.4",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "passlib[argon2]>=1.7.4",
    "psycopg[binary]>=3.2.12",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "uvicorn>=0.38.0",
]